Ensure citations containing ampersands are found
The underlying data source stores case body text as utf-8. Ampersands,
among other punctuation, is then stored in unicode, and not as its text
representation. To properly identify ampersands through regex, the regex
must identify "&amp;" and not "&". Fix this in the citation extractor.

Also add a test script to gitignore. In case this does not ultimately
resolve the issue, preserving my test script to return and delve further
at lower startup cost.